Drtii 1727 port dashboard tidying (#314)
* DRTII-1727 Refactor to db / json serialisation. Simplify

* DRTII-1727 Move tests into appropriate classes
