+++ shared/src/main/scala/uk/gov/homeoffice/drt/models/UserPreferences.scala
@@ -5,12 +5,23 @@ case class UserPreferences(userSelectedPlanningTimePeriod: Int,
                            showStaffingShiftView: Boolean,
                            desksAndQueuesIntervalMinutes: Int,
                            portDashboardIntervalMinutes: Map[String, Int],
-                           portDashboardTerminals: Map[String, Set[String]]) {
-  val serializedPortDashboardIntervalMinutes: String = portDashboardIntervalMinutes.map {
-    case (port, value) => s"$port:$value"
-  }.mkString(";")
+                           portDashboardTerminals: Map[String, Set[String]])
 
-  val serializedPortDashboardTerminals: String = portDashboardTerminals.map {
-    case (key, values) => s"$key:${values.mkString(",")}"
-  }.mkString(";")
+object UserPreferences {
+
+  def serializeMap[K, V](data: Map[K, V], valueToString: V => String): String = {
+    data.map { case (key, value) => s"$key:${valueToString(value)}" }.mkString(";")
+  }
+
+  def deserializeMap[V](data: Option[String], valueParser: String => V): Map[String, V] = {
+    data match {
+      case Some(s) if s.nonEmpty =>
+        s.split(";").map(_.split(":") match {
+          case Array(key, value) => key -> valueParser(value)
+          case _ => throw new IllegalArgumentException(s"Invalid format: $s")
+        }).toMap
+      case _ => Map.empty[String, V]
+    }
+  }
+
 }
